2-notebooks/2-agent_service/6-agents-az-functions/requirements.txt
azure-functions
azure-identity
azure-storage-queue